Add docs for shear and scale affine transformation Add test for the same.
