+++ src/mrsimulator/signal_processing/_base.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """The AbstractOperation class."""
+import numpy as np
 from mrsimulator.utils.parseable import Parseable
-
 
 __author__ = "Maxwell C. Venetos"
 __email__ = "maxvenetos@gmail.com"
@@ -20,6 +20,21 @@ class AbstractOperation(Parseable):
         if hasattr(self, "type"):
             my_dict["type"] = self.type
         return my_dict
+
+    @staticmethod
+    def _get_dv_indexes(indexes, n):
+        """Return a list of dependent variable indexes.
+
+        Args:
+            indexes: An interger, list of integers, or None indicating the dv indexes.
+            n: Total number of dependent variables in the CSDM object.
+        """
+        if indexes is None:
+            return np.arange(n)
+        if isinstance(indexes, int):
+            return [indexes]
+        if isinstance(indexes, (list, tuple)):
+            return np.asarray(indexes)
 
     @classmethod
     def parse_dict_with_units(cls, py_dict):
