+++ src/mrsimulator/signal_processing/tests/test_affine.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import csdmpy as cp
 import mrsimulator.signal_processing as sp
-import mrsimulator.signal_processing.affine_transformation as af
+import mrsimulator.signal_processing.affine as af
 import numpy as np
 
 # Creating a test CSDM object.
@@ -11,7 +11,9 @@ csdm_object = cp.CSDM(
     dependent_variables=[cp.as_dependent_variable(test_data)],
     dimensions=[
         cp.LinearDimension(count=40, increment="1 s", label="0", complex_fft=True),
-        cp.LinearDimension(count=40, increment="1 K", label="1", complex_fft=True),
+        cp.Dimension(
+            type="linear", count=40, increment="1 K", label="1", complex_fft=True
+        ),
     ],
 )
 
@@ -22,7 +24,7 @@ def test_shear_01():
     processor = sp.SignalProcessor(
         operations=[
             sp.IFFT(dim_index=1),
-            af.Shear(factor="-1 K/s", dim_index=1, normal=0),
+            af.Shear(factor="-1 K/s", dim_index=1, parallel=0),
             sp.FFT(dim_index=1),
         ]
     )
@@ -64,3 +66,47 @@ def test_shear_01():
     b1 = a1[1:-1] + 20
     b = np.append(a1, b1)
     assert np.allclose(index, [a, b])
+
+
+def test_serialization_and_parse():
+    processor = sp.SignalProcessor(
+        operations=[
+            sp.IFFT(dim_index=1),
+            af.Shear(factor="-1 K/s", dim_index=1, parallel=0),
+            sp.FFT(dim_index=1),
+        ]
+    )
+    serialize = processor.to_dict_with_units()
+
+    expected = {
+        "operations": [
+            {"dim_index": 1, "function": "IFFT"},
+            {
+                "dim_index": 1,
+                "factor": "-1.0 K / s",
+                "function": "affine",
+                "parallel": 0,
+                "type": "Shear",
+            },
+            {"dim_index": 1, "function": "FFT"},
+        ],
+    }
+
+    assert serialize == expected
+
+    recovered = sp.SignalProcessor.parse_dict_with_units(serialize)
+    assert recovered == processor
+
+
+def test_scale():
+    c_inc = csdm_object.dimensions[1].increment.value
+    c_off = csdm_object.dimensions[1].coordinates_offset.value
+
+    processor = sp.SignalProcessor(operations=[af.Scale(factor=2, dim_index=1)])
+    scaled_data = processor.apply_operations(data=csdm_object)
+
+    s_inc = scaled_data.dimensions[1].increment.value
+    s_off = scaled_data.dimensions[1].coordinates_offset.value
+
+    assert s_inc == 2 * c_inc
+    assert s_off == 2 * c_off
